ui/interactive: when abook quits, it does not rewrite the database anymore  *if* no change has been done to the items.
Running 2 abook instances concurrently is now less race-prone.
An indicator is added to the top line showing whether the database needs
saving on disk or not.

Warning: The indicator *is* introduced by the commit, but to stay safe, the
 change about optionnal database disk write is only effective for debug builds
 (= compiled using DEBUG = `make CFLAGS=-DDEBUG`)

Note: right now editing a field (even without changing to the contained string) is
 enough to be considered a "database change"